requirements.txt
django-allauth
django-avatar